feat: allow readonly input array
This method is meant for function arguments to be more versatile.
But the internal code wants to always have an array.

Using a readonly array as input allows for even more possible inputs.
This has the drawback of only a readonly output array.
Internally the array from input shouldnt be change anyway so this
shouldn't be a problem.

BREAKING CHANGE: returns only a readonly array